requirements.txt
alembic==1.13.3
bcrypt==4.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
cryptography==38.0.4
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysqlclient==2.2.4
passlib==1.7.4
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
flasgger==0.9.7.1
pyotp==2.9.0
